Move user info to MusicApp
- Move user info so that it isn't needed to put it in all components
- Create 'section-title' sass component to use it in all sections,
not only in the Home react component
- Run linter
- Remove unused `location` prop in Login and SignUp components
- use useUser in Artists component (making it a private route)
