+++ src/assets/sass/components/_section-title.scss
@@ -1,0 +1,7 @@
+.section-title {
+  margin-top: 20px;
+  margin-left: 15px;
+  margin-bottom: 0px;
+  font-weight: var(--font-weight-md);
+  font-size: 26px;
+}
