+++ src/assets/sass/layout/_main.scss
@@ -5,14 +5,6 @@
   line-height: 1.6;
 }
 
-h1.titulo_home {
-  margin-top: 20px;
-  margin-left: 15px;
-  margin-bottom: 0px;
-  font-weight: 400;
-  font-size: 26px;
-}
-
 h3.no_favoritos {
   margin-top: 20px;
   margin-left: 15px;
