+++ src/assets/sass/main.scss
@@ -10,3 +10,4 @@
 
 @import 'components/input';
 @import 'components/button';
+@import 'components/section-title';
