+++ src/assets/sass/pages/_artists.scss
@@ -3,7 +3,7 @@
   overflow: hidden;
 }
 
-.generoArtistas{
+.generoArtistas {
   margin-bottom: 10px;
 }
