+++ src/containers/Login/Login.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import AuthHelper from '../../components/AuthHelper';
 
-export default function Login({ location: { next } }) {
-  return <AuthHelper type="login" next={next} />;
+export default function Login() {
+  return <AuthHelper type="login" />;
 }
